Add human readable name to each run
Change the algorithms entry to runs, as it is not algorithms but a list
of runs, where each run is defined by the algorithm used and the
arguments given to the algorithm.

Instead of naming everything by the algorithm, which could clash and
does not tell us anything about the arguments passed in, we now allow
the benchmark specification to add custom name to each run.

Make the commandline outputs more user friendly and readable, providing
total number of steps in a group and incrementing it with each step.