Minor validation fixes
Turns out matlab and scipy do not use double to compute the results,
just our C++ implementation seems to be really sensitive to the width of
the datatype used. This should be investigated further in more detail in
the future.

Fix compute_valid_results.sh scripts not working with relative paths.
Now we make the paths absolute before passing the control off to matlab
or python.

CHanged underlying data type of result matrices in the matlab
implementation. THis has no influence over the computation, but seems
appropriate if we have input data in certain representation to have the
output data in the same representation.