Fix csv parsing bug
The CSV parsing code had an undefined behavior, where it was failing
only on a Release build and not Debug builds. This is due to the
std::string::substring method returning a copy of the substring, which
we then passed to boost::tokenizer, which retrieved iterators and
promptly left the string to deallocate. This only became a problem in
release build as only with this optimization level was the string
actually deallocated and the tokenizer started reading garbage.

This resulted in unknown escape sequence error being thrown by the
tokenizer.

The fix is to explciitly create a string_view of the header line and
create substring from this. We can allow tokenizer to create iterators
from the string_view before it is deallocated thanks to the following
from the C++ standard:

Note: Since the validity of iterators is not tied to the lifetime of an
object whose type models borrowed_­range, a function can accept
arguments of such a type by value and return iterators obtained from it
without danger of dangling.

The underlying array survives as long as the tokenizer, so it can use
the iterators without worry.