Add naive warp shuffle algorithm
First version of naive allgorithm using warp shuffle and caching in
registers.

The algorithm is derived from the original naive algorithm. Again each
thread computes a single shift.

The algorithm has two main branches, one for negative shifts and one for
non-negative shifts. They both work with the same idea, just the
non-negative one goes from left to right and the negative one from right
to left.

The non-negative part reads warp-size of the left-most overlapping
values from the right matrix. It then goes into a loop where it reads
the left-most warp-size of values from the left matrix and second
warp-size of values from the right matrix. The 2*warp-size of the values
from right matrix behave as a contiguous buffer which is shifted with
each step value down and the value from thread 0 is forgotten. THe
warp-size of values from the left matrix are then broadcasted, one with
each step, iterating from thread 0 to thread warp-size - 1. So
basically, each step we broadcast, compute the value and then shift the
big buffer one down. We do this for warp-size of steps and then load
another warp-size of values from right matrix into the top part of the
buffer, which will later be shifted down in the next cycle of steps, and
warp-size of values from left matrix which will be broadcasted in each
step in the next cycle.
This way we go through the whole row of the matrix, repeating it for
each row of the overlap.

The idea behind this is that because successive threads compute
successive offsets, if we read successive values from the right matrix
into each thread, they will all need to be multiplied with a single
value from the left matrix. In the next step, each thread needs value
the thread with rank +1 used in previous step, and needs to multiply it
with the value right next to the one broadcast before.

This is reflected in the algorithm for non-negative shifts directly. For
negative shifts, we do the same but instead shifting and broadcasting in
opposite order. Because we shifted both orders, it works the same.

The reason we split the branches is because it simplifies startup of the
algorithm. Due to the split, we can just load the first values used to
the corresponding threads as we can be sure that the matrices overlap in
the way we need.

The wrap-up of the algorithm, where the size of the matrix may not be
multiple of the warp-size is handled by bound checking every read from
the input matrices and if it is outside the bounds, we just return 0. As
the values are first used for multiplication and then for addition, the
zeroes have no effect on the result.

The result is mirrored in the x axis, so we may need to switch the usage
of right and left matrices, where we will be broadcasting from right
matrix and shifting values from left matrix. This will be added in a
future commit.