Profiling the naive orig and shuffle algorithsm
For smaller inputs up to 128, the number of active warps is too low and
we should increase it. This is true for both the orig and shuffle
algorithms.

One possibility is to use multiple threads per shift, splitting the rows
for the given shift between multiple threads. This distribution could be
dynamic based on the number of rows per given shift, as for example in
64x64 shift -63,-63 computes only a single row, whereas shift 0,0 goes
through all 64 rows of input.

The shuffle algorithm bottleneck is the LSU execution unit, which has
more than 90% usage for all input sizes. This is caused by each
multiplication and sum requiring 3 shuffles. We could reduce this
disparity by reading more right matrices into registers and computing
sums for multiple input right matrices at once. This would reduce the
ration to 4 shuffles per 2 multiplication additions, and this can be
repeated again and again.

These two optimizations can be combined.