Improve ration of shuffles to arithmetic operations
Add computation of the same shift of multiple right matrices by a single
thread. Instead of one broadcast shuffle and two down shuffles for each
multiply-add operation, we can now do N broadcast shuffles and two down
shuffles for N multiply-add operations, improving the ration of shuffles
to arithmetic operations.

This is to fix the bottleneck of LSU pipeline, which is the pipeline
implementing warp-shuffles. This pipeline has utilization over 90% in
all profiling sizes, so it is a hard bottleneck for this implementation.

The downside of this change is that with multiple right matrices
computed by a single thread, it limits occupancy as there will be N
times less threads for N right matrices computed by a single thread.

We will try to fix this in future commit, where we will change the
warp_shuffle algorithm to use multiple threads per shift, using the data
parallel nature of computing rows for given shift.

The change itself is simple. Instead of having a single variable for sum
and for thread_right, we make these local arrays with static size. Any
indexing into these MUST be unrollable and compile time expandable,
otherwise it is dynamic indexing and CUDA compiler will move these
arrays into local memory , which is basically automatically controlled
part of device memory, and it will as slow as reading from global
memory. This is why the NUM_RIGHTS must be a template argument and each
for cycle indexing the arrays must be bound by the constants 0 and
NUM_RIGHTS, so that the compiler can unroll it.

Add warp_shuffle one_to_many version. Other versions will be added once
we add the multiple threads per shift feature, so that we don't have to
change those.

Add n_to_m implementation using the original naive algorithm. I just now
realized that we can run the kernel multiple times, once for each left
matrix with all right matrices as right side arguments to
cross-correlation. This will not be optimal, as the kernel itself is not
optimized for this, but I am not changing it for now.