Add first simple implementation of warp per shift
Add first simple implementation of warp per shift. This implementation
uses the whole warp to compute a single output value for
cross-correlation, corresponding to a single shift.

The implementation abuses global memory even more than the original
implementation, as it uses 32 times more threads to compute the same
result, and as such reads the inputs 32 times more than the original.
This implementation is designed for the smallest inputs, as it optimizes
occupancy which is very low in the original implementation.

The implementation is optimized for low thread divergence, trading some
additional computations in each loop for basically no thread divergence.
We may want to test this in the future against a loop with more thread
divergence but simpler loop contents and as such faster loop execution.

In the future, we may reduce the global reads back by using shared
memory. The first parts of this implementation are prepared in the
ccn_shift_per_warp_shared_mem function and by the changes to
shared_mem_buffer.

Add alternative implementation for contiguous read to shared_mem_buffer.
Contiguous read has contiguous input and output. There are two
alternatives. First is to read warp_size amount of data and then
increment by block size. Second is to split the whole data range into
one chunk for each warp and read the whole contiguous chunkg by the
given warp. This should be better, as it optimizes the use of cache and
allows prefetch etc. We need to test this.

Add load_strided_chunks implementation, which is designed to read a
submatrix into a contiguous buffer. This will be useful for
shift_per_warp kernel, which will read the submatrices required by all
warps in given block and then do the computation only on the data in
shared memory, requiring much fewer reads from global memory.