Add work distribution to warp shuffle
Create a copy of the warp shuffle kernel, distributing computation of
single shift between multiple threads by computing separate rows.

The new copied kernel just does some more computation when computing the
indicies which it then passes on as arguments to the compute_matrices
device method. This method had a slight change added, where we add the
template argument ATOMIC which serves as a compile time switch between
atomic writes or non-atomic normal writes to global memory, so that we
preserve the behavior and throughput of the original warp shuffle
implementation.

Based on preliminary measurements, with correct configuration for row
distribution, this optimization is 15-25% faster than the original naive
implementation for sizes larger than 32x32. For sizes smaller than
these, the shift_per_warp implementation should be used, which is faster
than the original implementation roughly by the same percentage, even
without shared memory which will be added in future commits.