Add one_to_many implementations of warp_shuffle and shift_per_warp algoriths
Use the improvements to the shift_per_warp algorithm and warp_shuffle
algorithm, introduced in previous commits, which allow us to process
multiple right matrices.

The algorithms are even optimized to reuse the loaded data from left
matrix for these multiple right matrices.

Add no_distribution distribution, which should allow us to remove the
original implementations as it should behave the same. This needs to be
tested.

Fix warp_shuffle algorithm. Each CUDA block needs to belong to a matrix
group exclusively. Previous implementation had some blocks overlap
matrix group bounds and it expected the overlapping parts of warps to
compute different matrix group, which the code is not designed to do, so
we had first few values zero in the output as the search_size is one
below a multiple of two, and with block sizes of 32 each time one more
thread was overlapping a different matrix group, so first matrix group
was fine, second started with one 0 column, second with two 0 columns
etc.
The fix assigns each CUDA block to a matrix group and computes the
output offset based on the start of the matrix group, so whole matrix
group is now computed. We also needed to change how number of blocks to
be started is computed, to start enough blocks for all matrix groups.