Fix warp_per_shift shared mem rows
The algorithm was writing only the first matrix in the matrix group,
rest were left as 0. This was because it was computing warp out position
without any regard to matrix groups, so it only wrote one.

After fixing this, we got the first row in each output matrix zeroed
out. This was due to blocks overlapping consecutive matrices in matrix
groups, so the warps which are out of bounds and do only loading from
global to shared memory were writing their zero sums to the results at
row search_size.y, which was the first row of the following matrix.

We now check that the warp out pos is in range before writing outputs.