Remove unnecessary workers with no distribution
Due to fixed size of thread block, there will be more workers than
returned by num_workers. We want the workers to have no assigned work,
so we give these zero workers per the row.

Added explanation for lower precision with no distribution and the
original implementations with no work distribution.

The lower precision is caused by larger values of the intermediate
results to which we add each of the multiplication results. Due to the
larger values, the comparatively smaller results of the multiplication
lose more bits of precision the larger the value we are adding them to
is. This is inherent to the floating point format and there is nothing
we can do about it.

With work distribution, each intermediate sum is smaller, so less
precision loss on each addition. Afterwards all these intermediate sums,
which are now roughly the same magnitude, are added together atomically.
As they are roughly the same magnitude, there is almost no loss of
precision, so in the end we get more precise results.