Add run error recovery to benchmarking
When benchmarking, we don't want the whole benchmarking suite to stop on
the first failure of any run. We want to log the failure, store the
parameters which lead to the failure and continue benchmarking other
runs.

The functionality is added by this commit.

Instead of printing short message, return code, stdout and stderr, we
wrap these into ExecutionError class and raise the error. This error is
then caught by the group running the run, so it creates YAML
representation of the run, input and error and appends it to
failure_log.

There is one failure_log.yml per group, placed in the results directory.
The whole file containis one top level list of all errors, each item
containing the description of the run, error and inputs as described
above.

Add Group description to results directory, so that we have the
description of the runs stored even if the originla yaml file is
deleted. This description is per group, and describes the actual
configuration of the group and all the runs in the group as they are to
be executed.

Currently all serialization to YAML is done manually, by first
serializing to Dictionaries by to_dict methods and then combining these
dictionaries and writing them to yaml.

Changed the wierd no_overwrite_check flag to just overwrite flag, which
now deletes the previous group directory as whole and starts from clean
state.

Add handling when overwrite flag is not provided. The group is just
skipped instead of stopping the whole benchmark suite