Add FFT plan measurement
Measure cufftPlan calls to see how it behaves depending on input size
and data type.

From the results it seems that it is constant in regards of input size,
but for some reason plan for float takes 50% more time than already
pretty long plan time for double. Don't know why.

Rest of the computation behaves as expected, being much slower in double
than single, but due to the long plan times with small input sizes the
runtime of double is still shorter than the total runtime of single
variant.

Add jupyter notebooks for wapr_shuffle and fft.