Add superbuild
Add cmake superbuild, which builds the boost and nlohmann json
dependencies from source running the cross build.

The implementation is based on
https://github.com/Sarcasm/cmake-superbuild

Superbuild is optional, we still default to nomrla cmake build which
expects boost and nlohmann json to be installed system wide. But if we
are building somewhere where that is not possible, we can now run
superbuild using cmake -D USE_SUPERBUILD=ON, which first downloads and
builds boost, then downlaods and builds nlohmann json and only after
that builds our cross application which, regardless of what is installed
on the system, uses the boost and nlohmann json built previously.