Add free stage, prepare for inner measurement cycles, add optional output write
Add free measurement stage, where all allocated buffers and perpared
plans are freed. This is first to make the program more valid. Up until
now, we were just relying on the fact that everything is deallocated
once the whole process is killed. Now that we want to run measurements
without restarting process and keeping data loaded in memory, we will
need to properly deallocate everything so that we can allocate it again.

Add optional output write. As we don't want the outputs to hammer the IO
stack if we don't want the results anyway, such as with keep_output
false in benchmarks, change the --out option of the cross executable to
be optional. If it is not provided, the outputs are not written to file.
We should check that the compiler does not do any funny optimizations
with this change.

Separate alg total time measurement start from data loading. In the
inner measurement cycles, we may want to start even without loading
data.

Add timer reset to time measurement start. Another change preparing for
inner cycles.

Change benchmark.py script so that it does not create output directory
if we don't need it, such as in InternalRuns which don't need to keep
outputs. We still need outputs for ExternalRuns so that we can validate
if needed, but we then delete the files.