Propagate changes in measurement to benchmarking scripts
Add support for inner loops without data loading to benchmarking scripts

Add inner loop implementation to python and matlab cross corr
implementations

Add min_measure_seconds to benchmarking scripts. For now it is only
useful for C++/CUDA implementation, as python and matlab implementation
do not implement adaptive timing.

Add benchmark_type selection to benchmarking scripts.

Change validation for externalRun. Before, all results were computed
first and then validation ran on the results. Now only the inner_loops
are ran and validation runs on the results of the inner loops only. If
we are not keeping the results, they are then discarded and overwritten.

Fix validation results appending to a single file with inner and outer
loops. We need to print header only on the first inner loop of the first
outer loop, otherwise just append the result. The validation results
still read from standard output of the cross executable and then written
to a file after each outer loop.

Move global_config defaults to global config instead of checking if they
are None in the group parsing and setting them there. Now we can be sure
that the values will not be None even if not provided, simplifying the
code.

Make output of ExternalScript optional. This is passed from benchmarking
script to python or matlab and implemented there.

Remove the matlab cached/uncached benchmark. This can now be done using
inner loop of 1 for uncached and inner_loop of more for cached.

Change matlab and python implementations to output measurements more
closely resembling those measured by the cross executable. Remove
unnecessary categories measured. JUst added the Computation category
which will be the one we will be comparing, which measures the whole
computation without IO.

Fix the print_header_override in simple logger, which was not really an
override. Now it is.