Add n_to_m optimized warp shuffle implementation, with debug prints
Add implementation using warp shuffles which, in addition to computing
the shift with many right matrices, also computes shift with many left
matrices.

The principle is the same behind both optimizations, trying to improve
the ration of warp shuffles and arithmetic operations.

The changes to the original shuffling implementation, which computes the
same shift between one left matrix and multiple right matrices, are very
simple, basically just repeating what we do for the data from the right
matrices with the data from left matrices.

Each access to theleft matrix and data loaded from it is just wrapped in
a loop with static NUM_LEFTS bound, which should be unrolled in the
optimized code.

The thread blocks are now three dimensional, using the x axis to assign
left matrix group to blocks, together with distributing the shifts on
the x axis of the result matrix between blocks.

The y axis is used to assign right matrix group to each block.

The z axis is used to distribute rows of the result matrix between
workers.

Instead of having a giant switch which would dispatch all the
combinations of left and right matrix counts, we use recursive templated
dispatch methods which should basically generate a giant if else
statement. We should check that that is the case in the compiled code.

Move bound checked data loading methods, which are required by all
implementations using warp shuffle, into a separate header.

Add cuda streams to original naive implementation and original
implementation using warp shuffle, so that we can use them to implement
n_to_m and compare them with the optimized shuffling implementation.

Fix missing BenchmarkType::None dispatch in main.