Add adaptive measure to existing cross-correlation implementations
Add adaptive measurements to python and matlab implementations.

Pass the adaptive measurement limit through ExternalRun in benchmarking
to each of the scripts.

Change the shell adapters for each existing implementation to accept the
adaptive measurement limit.

Remove most measurement from the matlab implementation. As we only
really need the whole computation time, it is much easier to just
implement adaptive measurment of the whole algorithm instead of each
part separately.

All implementations of adaptive time measurement are mostly the same,
using the same principle with ratio and multiplication of iterations.
All support 0 time limit, which results in a single iteration
measurement. This will be most useful with matlab, which does crazy
amount of caching.

Fix ExternalRun names not following the convention of the internal runs
of empty four underscores if there are no arguments. As we do not yet
implement arguments for External runs, these will always be empty, but
must be present to enable easy processing in visualization.