Enclose contents local to the .cu files in anonymous namespace, format everything
Enclosed contents which should be private to each .cu file in anonymous
namespace, so that they do not pollute the global namespace and cannot
be linked to from other translation units by accident.

Removed the duplicate definition of warp_size constexpr and moved it
into the warp_size.hpp file. We do not use the builtin cuda warp size
variable as that represents read from builtin register in PTX, so it is
not constexpr.

Ran clang formatting on each cuda file so that they all match the style.
There should be no functional changes in this commit.

Fixed warning about unused parameters in the recursive dispatch
templates.