Revert unnecessary changes to the n_to_m shuffle
Revert unnecessary changes which made the n_to_m shuffle code hard to
read. THese were the merging of the two nested for loops, which as it
turns out was not the reason why nvcc was not optimizing the arrays into
registers. As the two nested for loops are more readable, we return them
back.